Move new components to correct folder
